+++ core/src/Control/Holdables/Holdable.java
@@ -1,0 +1,7 @@
+package Control.Holdables;
+
+import java.lang.reflect.Type;
+
+public interface Holdable {
+    public Type getSubclassType();
+}
