+++ core/src/Control/Holdables/Ingredients/Bun.java
@@ -1,0 +1,11 @@
+package Control.Holdables.Ingredients;
+
+import com.badlogic.gdx.graphics.Texture;
+import com.badlogic.gdx.graphics.g2d.Sprite;
+import com.badlogic.gdx.math.Vector2;
+
+public class Bun extends Ingredient{
+    public Bun(Vector2 position) {
+        super(position, new Vector2(20, 20), new Sprite(new Texture("badlogic.jpg")));
+    }
+}
