+++ core/src/Control/Holdables/Ingredients/Cookable/Cookable.java
@@ -1,0 +1,8 @@
+package Control.Holdables.Ingredients.Cookable;
+
+public interface Cookable {
+    /**
+     * Cooks the cookable Object
+     */
+    public void cook();
+}
