+++ core/src/Control/Holdables/Ingredients/Cuttable/Cuttable.java
@@ -1,0 +1,8 @@
+package Control.Holdables.Ingredients.Cuttable;
+
+public interface Cuttable {
+    /**
+     * Cuts the Cuttable Object
+     */
+    public void cut();
+}
