+++ core/src/Control/Holdables/Ingredients/Cuttable/Tomato.java
@@ -1,0 +1,22 @@
+package Control.Holdables.Ingredients.Cuttable;
+
+import Control.Holdables.Ingredients.Ingredient;
+import com.badlogic.gdx.graphics.Texture;
+import com.badlogic.gdx.graphics.g2d.Sprite;
+import com.badlogic.gdx.math.Vector2;
+
+public class Tomato extends Ingredient implements Cuttable {
+    private boolean isCut;
+    public Tomato(Vector2 position) {
+        super(position, new Vector2(20, 20), new Sprite(new Texture("badlogic.jpg")));
+    }
+
+    @Override
+    public void cut() {
+        isCut = true;
+    }
+
+    public boolean isCut() {
+        return isCut;
+    }
+}
