--- core/src/Model/GameState.java
@@ -1,12 +1,0 @@
-package Model;
-
-public class GameState {
-    public double timeLeft;
-    public int totalPay;
-    public int payGoal;
-
-    public GameState(double time, int payGoal) {
-        timeLeft = time;
-        this.payGoal = payGoal;
-    }
-}
