--- core/src/Model/Holdables/Holdable.java
@@ -1,9 +1,0 @@
-package Model.Holdables;
-
-import Model.WorldObject;
-
-import java.lang.reflect.Type;
-
-public interface Holdable {
-    public Type getSubclassType();
-}
