--- core/src/Model/Holdables/Ingredients/Bun.java
@@ -1,11 +1,0 @@
-package Model.Holdables.Ingredients;
-
-import com.badlogic.gdx.graphics.Texture;
-import com.badlogic.gdx.graphics.g2d.Sprite;
-import com.badlogic.gdx.math.Vector2;
-
-public class Bun extends Ingredient{
-    public Bun(Vector2 position) {
-        super(position, new Vector2(20, 20), new Sprite(new Texture("badlogic.jpg")));
-    }
-}
