--- core/src/Model/Holdables/Ingredients/Cookable/Cookable.java
@@ -1,8 +1,0 @@
-package Model.Holdables.Ingredients.Cookable;
-
-public interface Cookable {
-    /**
-     * Cooks the cookable Object
-     */
-    public void cook();
-}
