--- core/src/Model/Holdables/Ingredients/Cuttable/Cuttable.java
@@ -1,8 +1,0 @@
-package Model.Holdables.Ingredients.Cuttable;
-
-public interface Cuttable {
-    /**
-     * Cuts the Cuttable Object
-     */
-    public void cut();
-}
